Kludge for #5206 -- port of fix for #3536 to yaml
The internal format of parameters was changed between 0.25.x and 2.6.x, making
them incompatible when serialized.  This was fixed for PSON under ticket #3656
(by modifying the data as it was serialized) in:

  commit f66095d35bc5f9645eb19bbb8cefa342c0181d2d
and
  commit 2edf7fe968ac3d8af9865f65100940747c1fa894

but nothing was done about the other serialization formats.  This commit adds
a callback to zaml property serialization and uses it to route data through the
fixup method added in the commits above, thus fixing the problem "the same way"
as it was fixed for PSON under #3656.

It does nothing about marshal, mostly because I have not totaly gone over to
the dark side (yet).