Fix #5252 - line number mis-attribution during parsing
It is a resurgence of #2366 that appeared because of the commit
8971d8.

Before this commit, for associating documentation comments, we
were preferring line numbers coming from the parser currently reducing rule,
instead of the current lexer line number (which can be in advance
of several tokens due to the nature of LALR parsers).

We now merge the ast line number before fetching the comment from the
lexer.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>