Fix #4904 Mounts shouldn't remount unless they are ensure=>mounted
After we fixed issue #2730, it is now possible to manage an fstab entry
without asking puppet to try to call mount or unmount on that device.
That fix failed to address the "refresh" behavior of mounts.
We have changed "refresh" to only remount devices that are set to
"mounted", so users can truly manage fstab entries without having
puppet try to remount them.

Paired-With: Paul Berry <paul@puppetlabs.com>