Fix #5289 -- Bad copy/paste changes message on test failure
In my fix for #4894 (commit a097b939ab52bafb681cf7c5dcaf11717add07e6) I made
and tested the fix in one case and then copied most of it (all but a variable
initialization, Doh!) to two other locations.  This caused tests that would
have failed with a socket-in-use error to fail with a different error rather
than retrying.

Also fixed the spelling of "simultaneous."