(#5274) New comment property for the hosttype
When the parsefile provider for host parses /etc/hosts, it throws any
inlinecomment away. As a result they are also purged in /etc/hosts after
a puppetrun that detects a change. That could be dangerous because you
will lose information even about unmanaged resources.

So if you have something like

192.168.0.1 hostentry_not_managed_by_puppet # Important comment

in /etc/hosts the endresult will be

192.168.0.1\thostentry_not_managed_by_puppet

This patch introduces a new property "comment" for the host type. The
provider is nearly a complete rewrite and a lot shorter and hopefully
easier to understand.