(#5427) Using Propery::OrderedList for host_alias
This uses the propertyclass Puppet::Property::OrderedList to represent
the list of host_aliases. This lets us remove the in_sync, should_to_s
etc overrides.

In the provider class the list is represented by a string (=no array)
so there were a few changes necessary as well.

Because Puppet::Property::List uses the specified delimiter when
converting should values to strings, I changed the delimiter to a simple
space instead a tab. This keeps messages produced by puppet in a nice
format.

The tests had to be changed to work with the new behaviour of
host_aliases. There are a few additional tests as well.