(#5370) Made metrics and --summarize work with Puppet apply
Puppet apply and Puppet agent use different code paths to run the
transaction.  In the code path for Puppet apply, we weren't passing
the report to catalog.apply, and as a result catalog.apply was
creating its own report to store metrics in; this report was then
discarded.  As a result, reports generated using "puppet apply" had no
metrics, and the "--summarize" option didn't work at all.

Fixed by modifying the "puppet apply" code path to pass the report to
catalog.apply.  In the future it would be nice to refactor the two
code paths to be the same.

Paired-with: Jesse Wolfe <jesse@puppetlabs.com>