Maint: made upstart tests more robust.
The tests for the upstart provider were attempting to stub
Process::Status.exitstatus.  However, this doesn't work (presumably
because Process::Status is implemented in C).  As a result, the
upstart spec tests were failing if the most recent exit code was
nonzero.  Changed the tests so that instead of stubbing
Process::Status.exitstatus to return zero, they execute a command that
is known to succeed (`true`).

Paired-with: Jesse Wolfe <jesse@puppetlabs.com>