Maint: Make http handler code call the indirector through ".indirection"
In commit 71ecad9904c8c48c023e90e5fbea5b26b180c9cf we removed the
delegation from model class to indirection for the "find", "search",
"destroy", and "expire" methods.  When we did this we neglected to
modify http handler code to call the indirector directly.  This patch
makes the appropriate changes to http handler code.