(#5421) Link should autorequire target
When we manage a local link to a directory and the target directory is
managed by puppet as well, establish an autorequire. So if we have
something like

  file { '/foo': ensure => directory }
  file { '/link_to_foo': ensure => '/foo' }
  file { '/link_to_foo/bar': ensure => file }

we can ensure that puppet does not create dead links and does not try to
create '/link_to_foo/bar' before /foo is created.