(#5661) Creating types dont work with >1 namevar
The methods [] and []= of type.rb are handling :name in a special way.
When someone wants to access resource[:name] puppet tries to replace
:name with the name_var. If the type has more than one key_attribute
name_var is always returning false so we'll get the error

  Resource type <resourcetype> does not support parameter false

This patch doesnt try to substitute :name if we dont have a single
namevar, aka. we have more than one key_attribute