(#5662) Parsedfile doesnt work with mult keyattr
When one wants to use the parsedfile provider for a type with more than
one key_attribute (e.g. a type for entries in /etc/services with name
and protocol as key_attributes) the provider will not store all
key_attributes in property_hash and not all keyattributes will be
visible in the to_line function.

The create method of parsedfile will only put validproperties into the
propertyhash. As result :name and all the other key_attributes will not
be set.

In the flush method however the :name parameter is put in the
property_hash but the method does not handle other keyattributes.

This patch modifies flush to put all key_attributes into the property
hash
(Note: @resource.name is basically just an alias for @resource[:name])