(#5408) Reworked ResourceHarness so that code is clearer and all behaviors are tested
This patch removes the Puppet::Transaction::Change class and replaces
it with a method,
Puppet::Transaction::ResourceHarness#apply_parameter.  The new code is
shorter, more thoroughly unit tested, and addresses known bugs in
the interaction between auditing and performing changes.

This code does not address drawbacks in the report output (for example
a resource is still flagged as changed even if it merely contains
audit information); those will be addressed in a follow-up patch.