Base class for network device based providers
This is the common bits of all future network device providers
that are using prefetching/flushing to limit the number of
calls to the remote network device.

The idea is that we need one transaction to prefetch and one
to flush each instance.

Implementors needs to implement lookup which returns a hash
of the found entity, and flush to update the remote device.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>