maint: Work-around for a Mocha bug
Mocha is leaving behind a copy of stubbed methods in the singleton class of the
object on which the method is stubbed. So a later stub to any_instance of the
class isn't seen by the original object.

This change works around the problem by adding a follow-up "test" that removes
the method from the singleton class.

See: https://github.com/floehopper/mocha/issues#issue/20

Paired-With: Paul Berry