(#5710) Removed unnecessary calls to insync?
In the resource harness, we were calling insync? on all properties of
a resource, even if those properties weren't being managed.  This was
unsafe.  Changed to only call insync? on properties that are mentioned
in the manifest.

In addition, we discovered that the resource harness's computation of
desired_values was incorrect, and was consulting the current system
state rather than the desired state.  Since this hash was
(erroneously) only being consulted to see if it included :ensure, this
didn't cause any obvious bugs.  However it is now fixed.

Paired-with: Jesse Wolfe <jesse@puppetlabs.com>