(#5910) Improved logging when declared classes cannot be found:
Previously, when a class could not be found, it was displaying the same error message as when a resource type could not be found. This resulted in confusing error message: Invalid resource type class, when really it should display the name of the class that could not be found.

My patch changes the error message to:

Could not find declared class #{title}