(#5887) New type for handling solaris projects
Running Solaris you can create projects to group processes and do
resource management.  Every process that runs under a given project can
be limited in cpu resources, open files etc.  The projects are normaly
stored in /etc/project.  A project consists of an id, a name, a list of
users that are allowed to start processes under the given project, a list of
groups that are allowed to start processes under the given project and a
number of attributes for resource management.

Added a native puppet type to handle projects. Sample usage:

    project { 'proj1':
      projid           => '1000',
      users            => [ 'user1', 'user2' ],
      user_membership  => inclusive,
      groups           => [ 'group1', 'group2' ],
      group_membership => minimum,
      attributes       => [ 'task.max-lwps=(priv,100,deny)' ],
      ensure           => present,
    }