(#5479) Autorequire zfs filesystem when zone dataset is configured
 A zone dataset is just a zfs filesystem in the global zone.

This zfs filesystem needs to exist before it can be given to a zone as a
dataset.  It seemed to make sense to autorequire the zfs filesystem.  This
patch just autorequires the zfs filesystem which will be the dataset and let's
the zfs type manage autorequiring the parent zfs filesystems and zpool.

Reviewed-By: Daniel Pittman <daniel@puppetlabs.com>
Reviewed-By: Matt Robinson <matt@puppetlabs.com>