Feature #2597 -- nicer reporting of relationships.
Split out the reporting from a single line (often with literally hundreds or
thousands of items) into a multi-line report.  This is still nasty, but at
least it is easier to use as input to other systems.

This will also auto-join to a single line when sent to targets such as syslog
that do not approve of newlines in messages; this preserves the utility of the
message without needing to lose console utility.