(#4914) Prefetch mountstate
Add a method mountinstances that returns a list of currently mounted
filesystems and overwrite prefetch to update the :ensure state of a
mount resource to either
- :absent    => not in fstab and not mounted
- :unmounted => in fstab but not mounted
- :mounted   => in fstab and mounted
- :ghost     => not in fstab but mounted

This is just one step towards 4914. Next will be query the property_hash
when asking mounted? and update the insync? methods to handle the
different states.