(#4914) Query property_hash for mountstate
Change mounted? to query the property_hash so we don't have to run the
mountcommand for every specified mount but rely on the prefetched state

Also update the prefetched property_hash[:ensure] after mount or umount.
This is important if we query mounted? later (most obvious when refresh
is called. We dont want to remount a resource that was umounted before)