Fixed #6009 - nested member list vs directory service group provider
The Directory Service group (and user) provider behaves erratically if members
is a nested list; this happens with the following manifest:

    $r = ["root"]
    $a = ["daemon", "crc"]
    $n = ["nobody"]
    group { "testgroup":
      ensure => present,
      members => [$r, $a, $n]
    }

This resolves the issue by flattening the list at the time we are using it;
while a more general solution might be desirable this resolves the specific
issue cleanly enough.

Original patch from Clay Caviness.
Tests by Daniel Pittman <daniel@puppetlabs.com>

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>
Signed-off-by: James Turnbull <james@lovedthanlost.net>