(#5931) Prevent errors when calling insync? on audited properties
Created a method safe_insync? which first checks whether the property
has a "should" value of nil, and if so returns true.  Most insync?
methods were already doing this, but a few were not, leading to bugs
if a property was being audited but not set.

Types should continue to override the insync? method, but callers of
insync? should call safe_insync? instead.

Paired-with: Jesse Wolfe <jesse@puppetlabs.com>