Feature #2597 -- generate a DOT graph of cycles on request.
When the '--graph' option is specified, generate a new 'cycles.dot' file and
report the location of that to the user.  This contains only the cycles, in
dot format, allowing a visual representation of the cycle to be obtained
quickly.

This will include up to 10 paths through the cycle in the graph, and only one
in the display to the user, to reduce information overload.