(#4139) hook log autoflush into global defaults
We previously had an ordering dependency in the autoflush option, which was
statically read from defaults when the log destination was configured.

We add a hook in the defaults to update the log subsystem, which in turn
updates log destinations, when autoflush is changed.

This would work as desired:
  puppet agent --autoflush --logdest=file

This would not work, as autoflush would be false:
  puppet agent --logdest=file --autoflush

Now those changes propagate correctly.

Paired-with: matt@puppetlabs.com