(#5823) run mode can now be set dynamically...
Third party scripts, and complex command line tools, depend on being able to
configure the run_mode value at runtime, not just when they fire up.

For better or worse we used to allow this sort of thing to work, but stopped,
and we have no sane, safe and consensual alternative, so we broke a bunch of
client code.

This enables the feature again, but does not add any safety catch; you can now
happily slice off your own feet with this, if you really want to.