Fix #6267 - puppetdoc embedded links to puppet entities are not hyperlinked
Puppetdoc was relying on RDoc to provide the puppet entity (class, definition
node, etc...) hyperlinking in comments.
Unfortunately, RDoc was assuming namespaces are capitalized like Ruby
namespaces and that definition would use the # or . separator.

This change adds on top of RDoc puppet namespace format for classes and
definition.

This will make sure the comment hyperlinking will work as intented.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>