(#5977) fix spec test failure when new applications are introduced.
The test here was previously fragile, in that it would break when new
applications were introduced, and in that it depended on the order of items
returned from reading the directories on disk.

It is now insensitive to those changes, and still verifies that the results we
require occur, reducing long term maintenance cost.

Reviewed-by: James Turnbull <james@puppetlabs.com>