(#2) Should not assume interfaces have indirectors
The initial work assumed that all interfaces were just
skins on an indirected data type, but some interfaces will
be more abstract than that.

This commit removes that assumption by extracting all of
the indirector work into a new Indirector subclass of
Interface and then makes all of the new interfaces a subclass
of that rather than of Interface itself.