(#6335) Allow optional trailing comma in argument lists.
Support optional comma at the end arguments lists for parameterized
class and user defined resources. This makes it behave similar to
resource attributes.

The parser.rb file is generated with racc 1.4.6 against the grammer.ra
file and if this conflicts please remove parser.rb file and run make in
the directory to generate it.