(#1204) Move man generation task from install.rb to a rake task
Having this be commented-out code in the installer script doesn't really
make sense for our workflow; we want to be able to regenerate these manpages at will
and make sure they're always up-to-date. This will help us keep them in sync with the help text.

This commit also changes the ronn invocation to specify a manual name and organization.