(#6309) Ensure the correct device is mounted when managing mounts
Previously the mount type would only check if anything was mounted at the
desired point, when 'ensure => mounted' was specified.  Now we check not
only whether something is mounted at the desired point, but also that it
is the thing we wish to be mounted there.

There is also a chance that the mount point directory could be
"automagically" removed for us, when unmounting incorrect devices, so we
attempt to re-create the directory after unmounting to give the mount of
the correct device a better chance at succeeding.

Paired-with: Matt Robinson <matt@puppetlabs.com>
Paired-with: Nick Lewis <nick@puppetlabs.com>
Paired-with: Jesse Wolfe <jesse@puppetlabs.com>