(#5132) Provide a query REST interface for inventory
This REST interface returns a list of nodes that match a fact query.
Fact queries can use (in)equality testing as a string comparison, and >,
<, >=, <= numerical comparisons.  Multiple tests can be done as AND
comparisons, not OR.

The fact queries need to be prefixed by facts, and the comparisons other
than equality are specified with a .comparison_type after the fact name.

This will be better explained in the REST documentation on the website.

Searches that don't match anything now return empty array instead of a
404 error.

Conflicts:

	spec/spec_helper.rb