(#6395) Add extpuppet help, eval, and interfaces
This patch add three features to the envpuppet helper script.

First, if --help is passed as the first argument, a helpful message is
printed.

Second, if ENVPUPPET_BLEEDING=true when envpuppet is executed,
additional repositories are enabled.  The only addition at this time is
puppet-interfaces, which heavily modifies the behavior of Puppet itself
and may or may not be desired.

Third, if no arguments are passed to the script, bourne compatible shell
statements are printed to STDOUT which may be directly evaluated in the
current shell.  This is to help people who prefer to directly set
RUBYLIB and PATH in their development workflows, which I've noticed to
be the case for many people.

For example:

    eval $(envpuppet)

Signed-off-by: Jeff McCune <jeff@puppetlabs.com>
Reviewed-by: Jacob Helwig <jacob@puppetlabs.com>