(#6353) Restore the ability to store paths in the filebucket
Commit 2274d5104f6e413a2b8899a3c3111a17bbb2f4d7 optimized network
usage for the case where a file is already in the filebucket.
However, it took away the ability to store paths.

This change restores the ability to store paths while maintaining
optimal network usage for the case where the file is already in the
filebucket with the given path.  This is expected to be the most
common case.

Paired-with: Jesse Wolfe <jesse@puppetlabs.com>