(#6418) Recursive files shouldn't be audited
A vestigial codepath was accidentally made live again when 2.6.0's audit
parameter was added.  This patch removes that code.

As it's very difficult to write a meaningful unit test of a negative
case, a test will be added to the acceptance test project to confirm
before & after behavior for this fix.

Reviewed-By: Markus Roberts <markus@puppetlabs.com>